+++ ProteusMMX_382/ProteusMMX/ProteusMMX/Model/InventoryModel/Stockroom.cs
@@ -8,13 +8,12 @@ namespace ProteusMMX.Model.InventoryModel
 {
     public class Stockroom
     {
-        
         public int? StockroomID { get; set; }
         public string StockroomName { get; set; }
         public int? NumberOfParts { get; set; }
         public int? QuantityOnHand { get; set; }
-        public decimal? QuantityAllocated { get; set; }
-        public decimal? TotalQuantityAvailable { get; set; }
+        public int? QuantityAllocated { get; set; }
+        public int? TotalQuantityAvailable { get; set; }
         public decimal? TotalCostAmount { get; set; }
         public string TotalCostCurrency { get; set; }
         public string UserField1 { get; set; }
@@ -29,6 +28,3 @@ namespace ProteusMMX.Model.InventoryModel
         public string UserField10 { get; set; }
     }
 }
-    
-
-
